requirements.txt
black==21.11b1
eth-brownie>=1.17.2,<2.0.0
flake8==3.8.4
isort==5.7.0
brownie-token-tester>=0.2.2
